Pick an existing set of subsets if possible
Make AnalysisSelectionAndExecutionService package-private and make
determineAnalysisType() a static utility method.
