+++ gemma-core/src/test/java/ubic/gemma/core/analysis/expression/diff/LowVarianceDataTest.java
@@ -19,7 +19,6 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.springframework.beans.factory.annotation.Autowired;
 import ubic.basecode.util.FileTools;
-import ubic.gemma.core.analysis.expression.diff.DifferentialExpressionAnalyzerServiceImpl.AnalysisType;
 import ubic.gemma.core.analysis.service.ExpressionDataMatrixService;
 import ubic.gemma.core.datastructure.matrix.ExpressionDataDoubleMatrix;
 import ubic.gemma.core.loader.expression.geo.AbstractGeoServiceTest;
@@ -42,6 +41,7 @@ import java.util.Collection;
 import java.util.HashSet;
 
 import static org.junit.Assert.*;
+import static ubic.gemma.core.analysis.expression.diff.DiffExAnalyzerUtils.determineAnalysisType;
 
 /**
  * Test based on GSE19480, see bug 3177
@@ -151,8 +151,7 @@ public class LowVarianceDataTest extends AbstractGeoServiceTest {
     public void test() {
         ee = expressionExperimentService.thawLite( ee );
 
-        AnalysisType aa = analysisService
-                .determineAnalysis( ee, ee.getExperimentalDesign().getExperimentalFactors(), null, true );
+        AnalysisType aa = determineAnalysisType( ee, ee.getExperimentalDesign().getExperimentalFactors(), null, true );
 
         assertEquals( AnalysisType.TWO_WAY_ANOVA_NO_INTERACTION, aa );
